New way to open files in Linux
For Linux only:
Removed use of yad as it was not full featured enough (no thumbnail
view, and did not allow multiselect with custom order preserved).
Created a new python application which:
Opens a Dolphin(flatpak only) browser to allow the user to copy the
files they want to import.
Thumbnails in the Python app allow the user to drag to rearrange the
images as they wish before accepting the files into the blog.
New requirement for installing Dolphin as a flatpak.  Dolphin is the
only file browser I found which allows clicking a + or - on the file
thumbnails to multi-select like windows allows.