Improved Linux File Opening
- Added a Python application "PasteNsort" which allows the user to paste
a list of files from a file browser (dolphin is opened) which will be
shown as thumbnails.  The user can then drag the thumbnails to sort them
before accepting the file list in proper order into the Blog.  This
required adding the file "video.png' which is shown as the thumbnail
when a video is loaded.
- Adjust button name for checking resources.
- Adjust button for media cleanup so it greys out when clicked.
- Update Linux ImageMagick to use new version (same as Windows).
- Added Media Cleanup console log messages